+++ ChangeLog.markdown
@@ -6,4 +6,5 @@ This is the list of changes for the PHPUnit_Selenium 1.1 release series.
 PHPUnit_Selenium 1.1.0
 ----------------------
 
+* Implemented GH-31: Skip tests when connection to Selenium RC server is not possible.
 * Fixed HTTP deadlock with Selenium RC server.